pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "f1rl"
version = "0.1.0"
description = "Professional-grade multi-agent Formula 1 simulation with reinforcement learning."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Lucas" }
]
dependencies = [
  "numpy>=1.26",
  "PyYAML>=6.0",
  "torch>=2.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0"
]

[project.scripts]
f1rl-train = "f1rl.cli.train:main"
f1rl-evaluate = "f1rl.cli.evaluate:main"
f1rl-simulate = "f1rl.cli.simulate:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
